requirements.txt
fastapi==0.115.2
streamlit==1.45.0
uvicorn==0.29.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
faiss-cpu==1.11.0
transformers==4.51.3
peft==0.13.2
cryptography==44.0.3
torch==2.7.0
sentence-transformers==4.1.0
PyMuPDF==1.23.22